Improving the update logic for TileEntitySprinkler
This commit replaces the BlockRange.stream() approach with a per-column
search so that obstructions (anything except air or plants) block the
effect. The search is also stretched out over multiple updates, so only
one column is done per tick.

Additionally, the configuration descriptions have been updated so that
they fit with the new logic. The growth interval is now the minimum tick
count before starting the loop again. It will not take less than the
number of columns in the cuboid. The default radius is 3, so that's a
49 ticks minimum.

Finally, while the water usage is still mB/second (20 ticks), it now
handles all values correctly (instead of multiples of twenty only). As
a result, for a rate of R/second, the first (20 - R%20) ticks will
consume (R/20) mB, and the last (R%20) ticks will consume (R/20 + 1).
Could be smoother, but this is fast and pretty good. Worth mentioning
is that the sprinkler will consume water continuously, even during the
downtime between irrigation passes.

This patch also cleanly divides the irrigation process into a simple
and straight-forward method for updating the sprinkler's state
variables, and a method for handling an arbitrary column of irrigation.
They are glued together with a helper method that fills in the
necessary parameters.

The NBT read and write methods now save the value of the buffer, in
addition to the new persistent state variables. The read code now also
relies on the documented default values that are returned when the
requested key does not exist, greatly simplifying that section.

Lastly, most of the code introduced by this patch has documentation
included with it.