Improving the update/irrigation logic for TileEntitySprinkler.java
This commit replaces the BlockRange.stream() approach with a per-column search
so that obstructions (anything except air or plants) block the effect. The
search is also stretched out over multiple updates, so only one column is done
per tick.

Additionally, the configuration descriptions have been updated so that they
fit with the new logic. The growth interval is now the minimum tick count
before starting the loop again. It will not take less than the number of
columns in the cuboid. The default radius is 3, so that's a 49 ticks minimum.

Finally, while the water usage is still mB/second (20 ticks), it now handles
all values correctly (instead of multiples of twenty only). As a result, for
a rate of R/second, the first (20 - R%20) ticks will consume (R/20) mB, and
the last (R%20) ticks will consume (R/20 + 1). Could be smoother, but this
is fast and pretty good. Worth mentioning is that the sprinkler will consume
water continuously, even during the downtime between irrigation passes.

An indirect result of these logic changes is that the read/writeTileNBT
methods plus the reference/AgriNBT.java class also got updated to save and
restore these new state variables as needed.

Edit Comment: Since the final product entirely* replaced all the changes of
the original patch, it made more sense to rebase this commit alone on to the
master branch. *-> Also fixes the TileEntityChannel.java typo in the offset,
which caused channels to check above, instead of below, themselves for any
sprinklers to fill with water.