Make onBlockActivated return true for fertilizers
In order to prevent PlayerControllerMP.processRightClickBlock from
trying to pass the click to other methods, onBlockActivated has to
return true. In the case of fertilizers, we want to block vanilla
behavior, especially when agricraft says that the fertilizer shouldn't
cause an update. Previously, this method would return true only if
there was a plant or if fertilizer mutations were enabled, and otherwise
return false and thereby permit vanilla code to take over.

This *directly* fixes issue #1018.

Stopping the right click process early saves time, and ensures that
agricraft fertilizer behavior maintains control.