Make onBlockActivated return true for fertilizers
In order to prevent PlayerControllerMP.processRightClickBlock from
trying to pass the click to other methods, onBlockActivated has to
return true. In the case of fertilizers, we want to block vanilla
behavior, especially when agricraft says that the fertilizer shouldn't
cause an update. Previously, this method would return true only if
there was a plant or if fertilizer mutations were enabled.

This *directly* fixes issue #1018.

Stopping the right click process early saves time, ensures that
agricraft fertilizer behavior maintains control, and also avoids any
problems with BlockCrop.canGrow(). (That was also fixed separately.)