Fix creating BlockCondition with the right amount
This depends on a patch to AgriCore that adds an accessor method for the
protected AgriCondition#amount field.

In JsonPlant#initGrowthRequirementJSON, BlockConditions are created from
the plant's AgriConditions. The amount of blocks required is derived
from the size of the ItemStack in the FuzzyStack parameter. The bug is
that AgriCondition#toStack(FuzzyStack), aka AgriStack#toStack(token),
sets the size to 1, instead of the value of the amount field.

This fix uses the added AgriCondition#getAmount accessor to instead call
toStack(token, amount). Then when the constructor for BlockCondition
sets the value for amount, the stack will provide the correct value.
# Please enter the commit message for your changes. Lines starting
# with '#' will be ignored, and an empty message aborts the commit.
# On branch patch-blockcondition-amount
# Changes to be committed:
#	modified:   lib/AgriCore
#	modified:   src/main/java/com/infinityraider/agricraft/core/JsonPlant.java
#
# Changes not staged for commit:
#	modified:   src/main/java/com/infinityraider/agricraft/utility/BaseIcons.java
#