Add new DebugMode for testing IGrowable blocks
This mode mimics the logic that is in ItemDye#applyBonemeal, but also
provides feedback via chat about each step as it happens.

It will call IGrowable#canGrow, #canUseBonemeal, and then #grow. If
either method returns false, then the attempt stops at that point.

Currently, it generates a puff of smoke after #grow is called, as extra
feedback. That could be a config option instead. Or the chat feedback
could be more tidy and then be considered sufficient.