Rewrite BlockCrop IGrowable methods, add a config
This commit has essentially four effects.

1) It fixes canUseBonemeal so that the IGrowable interface isn't stuck
  in an inbetween state of allowing vanilla bonemeal (and things that
  mimic it) to be used and consumed, but not have any actual affect on
  the crop.

2) It switches the three methods to wrap around AgriCraft's fertilizer
  system, instead of using custom tests and logic. Now the interface
  will respect the same rules as bonemeal. For example, it is blocked
  by plants that forbid fertilizers.

3) It adds a new boolean option to the farming catagory that outright
  enables and disables the interface. It defaults to enabled, which
  means that AgriCraft should now be more compatible with mods that
  speed up plant growth. But it can be disabled for purer gameplay.

4) The documentation for all of the updated methods has been greatly
  expanded. Requirements and conditions are detailed. The behavior and
  usage of the interface in vanilla is described.

The rewrite adds a private method that handles getting the correct
references to the crop and the bonemeal adapter, and using them to make
the necessary calls. This way code isn't duplicated across the three
interface methods. It handles both the checks and the growth calling,
so that the look ups don't need to be done twice. It also protects
against NPEs.

The IGrowable interface can be difficult to test in regular AgriCraft,
but a separate patch adds a new mode to the debugger for that.