Update TileEntityCrop#IAgriFertilizable (and docs)
This update of TileEntityCrop#acceptsFertilizer and #onApplyFertilizer
adds some missing features, while reducing the code complexity and
redundancy. It also expands the Javadoc for IAgriFertilizable,
specifying exactly what the different permutations of input conditions
are, and what the expected behavior is.

Firstly, fertilizers can now spawn weeds, as was requested by RlonRyan
and InfinityRider. Fertilizers are still controlled by the
fertilizerMutation config and the isFertilizable/canBonemeal property of
lants. Furthermore, other than these two exceptions, fertilizers now use
the same code and do the same things that normal random ticks do. Note
that this second part is actually only a small change: if fertilizers
can cause Growth, Cross-Overs, and Spawning, then the only thing that
was missing was making mature plants try to Spread.

This updated behavior allows the #onApplyFertilizer method to be
simplified, and for the specialized logic within it to be removed.
Previously it had been partially (but not entirely and cleanly) separate
and parallel to the #onGrowthTick method. Now all the subroutine
branching occurs in #onGrowthTick, and #onApplyFertilizer simply wraps
around it to filter out the cases where fertilizers are blocked.

This somewhat alleviates the bug where bonemeal is incorrectly consumed
(#1018), since #acceptsFertilizer returns false in fewer situations. The
full fix for that bug is in a separate commit.